+++ CHANGELOG.md
@@ -5,7 +5,9 @@
   * Generating `variables.tf` for capability modules.
   * Generating `capabilities.tf`, `capabilities.tf.tmpl`, `outputs.tf` for app modules.
 * Fixed `--connection` flags in `blocks new` command.
-* Updated `workspaces select` command to generate `capabilities.tf` from `capabilities.tf.tmpl`.
+* Updated `workspaces select` command:
+  * Generating `capabilities.tf` from `capabilities.tf.tmpl`.
+  * If new `ns_connection` exist locally, will prompt user to select a target for the connection.
 
 # 0.0.46 (Feb 24, 2022)
 * Updated CLI to utilize new stack-based API endpoints in Nullstone API.